requirements.txt
fastapi
uvicorn
sympy
openai
python-dotenv